WASM read/write memory separated into wasm_memory
